+++ tests/PsKyc/Service/NotificationServiceTest.php
@@ -14,7 +14,7 @@ namespace Tests\PsKyc\Service;
 use Mockery\Adapter\Phpunit\MockeryTestCase;
 use PrestaShop\Module\Pskyc\Service\NotificationService;
 use Symfony\Component\Templating\EngineInterface;
-use Symfony\Component\Translation\TranslatorInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class NotificationServiceTest extends MockeryTestCase
 {